requirements.txt
Django==2.2.16
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
